Added toggleGroup types
